Deep Live/requirements.txt
numpy>=1.23.5,<2
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.16.0
insightface==0.7.3
psutil==5.9.8
tk==0.1.0
customtkinter==5.2.2
pillow==11.1.0
onnxruntime-gpu==1.17
tensorflow
opennsfw2==0.10.2
protobuf==4.23.2
